Raise if JS.dispatch detail is not a map (#4062)
Today I discovered a bug in a LiveView application where a string was
passed as the detail option to JS.dispatch instead of a map (as
documented).

It turns out it accidentally worked with the default `esbuild` config.

The bug remained unnoticed for a long time because by default `app.js`
is not compiled in JavaScript strict mode. When the config was changed
to output ESM modules (which are always in strict mode), the error
surfaced.

The exact line of code in Phoenix LiveView client code where the error is raised:
https://github.com/phoenixframework/phoenix_live_view/blob/98088bc7458d1821d9d9bb38caa3145a0b93ee9c/assets/js/phoenix_live_view/js.js#L77
Note that line has been there for many years.

This is the explanation of the error in MDN:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Cant_assign_to_property

LiveView code is correct and follows the documented API. Since we
already do validations and raise on other invalid inputs, we can prevent
future mistakes by ensuring that the detail option is a map.