Make bug report template easier to fill (#4076)
* Make bug report template easier to fill

This is an attempt to make it easier for users to fill out the bug
report template by requiring less edits to copy-paste the relevant
information.

Hopefully, this translates into better reports that deviate less from
the template and less friction for maintainers to triage issues.

* Drop node_modules questions from bug report template

Based on feedback from Steffen.

Probably the info is not very relevant in most cases.