Support pushing JS commands via events (#4060)
* Add JS.to_encodable/1 function for pushing JS commands via events

The new `JS.to_encodable/1` function transforms the
`Phoenix.LiveView.JS` struct into an opaque JSON-serializable value,
such that it can be transmitted outside of HEEx templates.

See discussion https://elixirforum.com/t/make-js-t-a-public-data-structure-or-json-serializable/53870?u=rhcarvalho.

* Conditionally implement Jason.Encoder and JSON.Encoder for Phoenix.LiveView.JS

Makes it more convenient to use JS commands in `push_event` payloads.

* Reword `JS.to_encodable/1` doc

- No need to link to hexdocs.pm, ExDoc will do that automatically
- Use the same wording as Phoenix, "JSON library" instead of "JSON encoder"
- More concise wording

* Add to_encodable/1 tests

* Add Jason.Encoder and JSON.Encoder tests

* Simplify js.exec test

1. Remove done callback, which is meant to be used with
   callbacks/asynchronous code.

https://jestjs.io/docs/asynchronous#callbacks

2. Use jest.runAllTimers(), reflecting typical usage when we don't care
   about intermediate timer states.

https://jestjs.io/docs/jest-object#jestrunalltimers
https://jestjs.io/docs/jest-object#jestadvancetimersbytimemstorun

* Add JS test coverage for execJS, js().exec and hook.js().exec

* Conditionally define JSON.Encoder test

To avoid warnings on older Elixir versions where JSON.Encoder is not
available.

When the JSON module is not available, we define a skipped test with
the same name for visibility in test reports.